requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
dnspython==2.2.1
email-validator==1.3.0
fhir.resources==6.4.0
fhirpy==1.3.0
frozenlist==1.3.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
marshmallow==3.18.0
multidict==6.0.2
numpy==1.23.2
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
pyserial==3.5
pysmb==1.2.8
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scapy==2.4.5
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
